requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-modeltranslation==0.15
django-recaptcha3==0.2.0
django-silk==4.0.1
gprof2dot==2019.11.30
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.2.0
pycodestyle==2.5.0
Pygments==2.6.1
python-dateutil==2.8.1
python-gettext==4.0
python3-openid==3.1.0
pytils==0.3
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
